# perf(api): use QuerySet.values to improve performance