prefetch_related is _slow_. Using QuerySet.values speeds things up by not creating model objects and avoiding a lot of code. 1162ms to 186ms.



**Before**
1162 ms
1239 ms
1144 ms

**After**
202 ms
191 ms
186 ms